+++ source-code/clj/server_extensions/storage/media_browser/mutations.clj
@@ -213,34 +213,19 @@
 ;; ----------------------------------------------------------------------------
 ;; ----------------------------------------------------------------------------
 
-(defn undo-duplicate-item-f
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  [env {:keys [item parent-id] :as mutation-props}])
-  ;(when-let [media-item (engine/get-item env (:media/id item))]
-  ;          (engine/update-path-directories! env           media-item +)
-  ;          (engine/attach-item!             env parent-id media-item)
-  ;          (return media-item)}])
-
-(defn undo-duplicate-items-f
-  ; WARNING! NON-PUBLIC! DO NOT USE!
-  [env {:keys [items parent-id]}])
-  ;(letfn [(f [result item]
-  ;           (conj result (undo-delete-item-f env {:item item :parent-id parent-id}))
-  ;       (reduce f [] items)}])
-
 (defmutation undo-duplicate-item!
              ; WARNING! NON-PUBLIC! DO NOT USE!
-             [env {:keys [item]}]
-             {::pathom.co/op-name 'storage.media-browser/undo-duplicate-item!})
-             ;(let [parent-id (item-browser/item->parent-id :storage :media item)]
-              ;    (undo-delete-item-f env {:item item :parent-id parent-id})})
+             [env {:keys [item-id]}]
+             {::pathom.co/op-name 'storage.media-browser/undo-duplicate-item!}
+             (let [parent-id (item-browser/item-id->parent-id :storage :media item-id)]
+                  (delete-item-temporary-f env {:item-id item-id :parent-id parent-id})))
 
 (defmutation undo-duplicate-items!
              ; WARNING! NON-PUBLIC! DO NOT USE!
-             [env {:keys [items]}]
-             {::pathom.co/op-name 'storage.media-lister/undo-duplicate-items!})
-             ;(let [parent-id (item-browser/item->parent-id :storage :media (first items))]
-              ;    (undo-delete-items-f env {:items items :parent-id parent-id})})
+             [env {:keys [item-ids]}]
+             {::pathom.co/op-name 'storage.media-lister/undo-duplicate-items!}
+             (let [parent-id (item-browser/item-id->parent-id :storage :media (first item-ids))]
+                  (delete-items-temporary-f env {:item-ids item-ids :parent-id parent-id})))