+++ source-code/clj/x/boot_loader/events.clj
@@ -1,0 +1,28 @@
+
+;; -- Namespace ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(ns x.boot-loader.events
+    (:require [x.server-developer.api]
+              [x.server-dictionary.api]
+              [x.server-environment.api]
+              [x.server-router.api]
+              [x.server-views.api]
+              [x.app-details     :as details]
+              [x.server-core.api :as a :refer [r]]
+              [x.server-db.api   :as db]))
+
+
+
+;; -- DB events ---------------------------------------------------------------
+;; ----------------------------------------------------------------------------
+
+(defn- store-server-props!
+  ; WARNING! NON-PUBLIC! DO NOT USE!
+  ;
+  ; @param (map) server-props
+  ;
+  ; @return (map)
+  [db [_ server-props]]
+  ; A szerver indítási paramétereinek eltárolása
+  (assoc-in db [:boot-loader :server-handler/meta-items]))